Bunting should only appear on specified days
Previously we had a problem whereby we would need to manually remove
view code to show the bunting when specific bank holidays were
occurring. This was purely out of respect, e.g. on Orangemen's Day.

This commit allows us to check whether a given day is allowed to show
bunting by seeing if it's within an allowed block. To simplify this
process, using the holiday titles rather than dates.