Adding a (hidden) date allows spreadsheet programs to parse dates corrently, when copy/pasting
For example, copying a table into excel or google drive, if the value of the date cell
is '2 April' then it will be detected as a date, but the year will default to the current.

For past/future years this results in the wrong year. If we provide a visually and AT
hidden year for this case then the date will be parsed entirely correctly.

This is doubly-hidden content, but as it's entirely progrmatic its unlikely to go stale,
as static copy tends to.

The non-breaking space is required to keep the right formating when copied/pasted.
With a regular space it would copy as D M Y, but be treated by google drive as D MY.

@edds suggested this as the reason:

> I would hypothesis that the missing space is due to the visuallyhidden element having position: absolute;. The position will give the element a display value of block. My theory is that the browser ignores the standard whitespace before the element as block level elements aren't effected by whitespace. Adding the &nbsp; will force the space to remain. Though as I say that is just a hypothesis and could be wrong.