Remove CONTRIBUTING.md
CONTRIBUTING.md mentions Travis and is currently out of date. Instead
of trying to keep this up-to-date we might be better off by trusting
that developers on the repo are familiar with the GDS/GOV.UK style
guides. Having a specific contributing file implies that there's
something different in this repo - which there is not.