Inline the client for panopticon registering
As pointed out by @alext, the instantation of the panopticon client
should be done in the rake task because it is a special type
(`GdsApi::Panopticon::Registerer`, not `GdsApi::Panopticon`) and will
only be used in this rake task.

This also fixes problems with autoloading of the `GdsApi` classes in
production.