Disable SASS linter for non-trivial violations
After discussion it's been decided to disable the linter for specific cases where the violations are not trivial. The app's front-end code has not been updated in a while and needs more general updating.
Leaving the linter disabled in these cases will serve as a clear sign post for required action.