Upgrade rails gem version to 4.2.7.1 (CVE-2016-6316)
As a result of the XSS vulnerability in Action View, text that pass as HTML safe can't have quotes
escaped when used as attribute values in tag  helpers.

This vulnerability has been defined as  CVE identifier CVE-2016-6316. See https://groups.google.com/forum/#!topic/rubyonrails-security/I-VWr034ouk]

This commit updates the rails version to 4.2.7.1 in order to benefit from the release patch that
caters for the aforementioned vulnerability.