Use tinsel automatically at Christmas and New Year
For the past few years bob and I have had a little Christmas tradition
of devopsing a change in production to use the tinsel bunting on the
bank holidays page over Christmas and New Year.

As part of our valiant quest to automate ourselves out of a job,
it turns out that these days computers can even do that for us.

Progress, hey?