Add heroku preview deployment script
To get a working preview of changes before merging
into master, we can deploy the branch to Heroku
using this script. It is a modified version of the
same script as used in

https://github.com/alphagov/smart-answers/blob/master/startup_heroku.sh

Readme has instructions as taken from

https://github.com/alphagov/smart-answers/blob/master/doc/factcheck.md

Heroku only supports reading the ruby version from
the Gemfile. It otherwise defaults to an older version of ruby,
that is incompatible with some of the gems being used.