Fix test
Upgrading to Rails 5 provokes this error when stubbing the render method:

```
Error:
CalendarControllerTest#test_: GET 'calendar' HTML request (no format) should load the calendar and assign it to @calendar. :
ActionController::UnknownFormat: CalendarController#calendar is missing a template for this request format and variant.

request.formats: ["text/html"]
request.variant: []

NOTE! For XHR/Ajax or API requests, this action would normally respond with 204 No Content: an empty white screen. Since you're loading it in a web browser, we assume that you expected to actually render a template, not nothing, so we're showing an error to be extra-clear. If you expect 204 No Content, carry on. That's what you'll get from an XHR or API request. Give it a shot.
    test/functional/calendar_controller_test.rb:18:in `block (3 levels) in <class:CalendarControllerTest>'
```

It's a bit nicer to test the outcome rather than inner code workings anyway, so we'll tweak the test.