Add a failing test for tinsel not being used in early December
2nd December 2019 is a substitute bank holiday in Scotland for St Andrew's
Day. The current implementation of `Calendar#bunting_styles` uses tinsel for
all bank holidays with `"bunting": true` in December and January, but it
should probably only do that for the ones for Christmas and New Year.

There were substitute days for St Andrew's Day in 2013 and 2014 as well. I've
used one of those in the test because for some reason the test isn't failing
today when I use the 2019 date - maybe something to do with Timecop?