Remove bank holidays for 2012, 2013 and 2014
We’ve had a chat about the past bank holiday dates in content support
and think it would be good to remove the 2012, 2013 and 2014 dates
(under Eng and Wales, Scotland and NI). We don’t know of any user needs
for keeping them, but don’t want to strip them back too far unless
there is a user need to see the previous few years’ bank holidays
that we’re not aware of.