Make the event UIDs for the ical files unique
We used to use indices to generate UIDs. This meant that when we deleted
some past bank holidays back in December 2019, the indices pointed to
different events.

Using a combination of event title and event date instead of indices to
create event UIDs means that these will remain unique even when we'll
delete other past bank holidays in the future.

We can use event title and event date to make the UIDs uniques because
there are never going to be two bank holidays with the same name on the
same day.

More info and background can be found here:
- [Commit that deleted old bank holidays](https://github.com/alphagov/calendars/commit/c28e7b570c39a10f60d2ed13b85d477b9a60159e#diff-1569053def8d48383b565a3d2d505529)
- [Issue raised by a member of the public](https://github.com/alphagov/calendars/issues/753)

Resolves #753

Trello card: https://trello.com/c/4MuxUJsy/1715-2-fix-the-event-uids-for-the-ical-files-provided-by-the-calendars-app